refactor(inbox): extract the triage/actions engine into inbox-engine.ts
Second of three behavior-preserving slices. Moves the triage + action-execution +
learning-extraction engine out of inbox.ts into inbox-engine.ts:
runTriageAgent, runProposedAction, runDispatchedAgentToTerminal, awaitRunTerminal,
maybeRefireTriage, maybeExtractLearning, maybeCommitBuiltAgent, the agent-editor/builder
auto-propose helpers, the crash-retry helpers, and the action-lifecycle counters — plus
the engine constants. The runTriageAgent <-> runProposedAction/maybeRefireTriage cycle
stays entirely inside this one module, so no cross-file import cycle.

inbox.ts is now just the 20 route handlers + handleLearningDecision + router wiring
(925 lines, down from 2249; inbox-engine.ts is 1355). Routes import the engine entry
points; test-imported symbols stay reachable via re-export shims.

Cleaned up a wart from the move: TRIAGE_AGENT_ID now lives in inbox-shared.ts (the leaf)
instead of the engine, so inbox-catalog.ts no longer imports from the router file — the
module graph is now strictly acyclic (shared <- catalog/plan/widgets <- engine <- routes).

Pure cut-and-paste, no function->const conversions, no logic changes. Clean build; full
dashboard+core suite unchanged at 2018 pass / 3 skip; live routes smoke-tested
(GET /inbox, respond->triage, fragment, action proposal).
